requirements.txt
aio-pika==9.5.8
aiormq==6.9.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
click==8.3.1
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
idna==3.11
multidict==6.7.0
pamqp==3.3.0
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
yarl==1.22.0
